usb: replace USB_OTG dependency and modified defconfig
The earlier commit to remove the unnecessary USB_OTG dependency broke
the USB Ethernet gadget support on Smasher.  Due to an issue with
Kconfig and building the CONFIG_USB_ETH option as built-in rather than
as a module, it is not possible to fix up that commit to work properly.
Reverting that commit to put things back the way they were, and filed
CAR 470001 to address the Kconfig issue in a future release.

Signed-off-by: Ben Shelton <ben.shelton@ni.com>
Acked-by: Joshua Hernstrom <joshua.hernstrom@ni.com>