8250_pnp: Support dual-mode UART through PMR
Some NI targets can have a NI 16550 UART which can be connected to
either a RS-232 transceiver or a RS-485 transceiver. This dual-mode
support is implemented through the Port Mode Register (PMR). This
change will add support for the 8250 PNP driver to read the PMR and
based on the mode, will register the appropriate driver.

Signed-off-by: Karthik Manamcheri <karthik.manamcheri@ni.com>
Acked-by: Jaeden Amero <jaeden.amero@ni.com>
Acked-by: Justin Tang <justin.tang@ni.com>