igb: Reset TXPBSIZE to default value
The TXPBSIZE register of the i210 resets to its default value only
at power-on. It doesn't reset if you reboot the system, only if you
pull power. If something (another driver, another OS, etc.) modifies
this register from its default value, the igb driver doesn't function
correctly. It detects a hang of the transmitter and continuously resets
the adapter. Here we set this value to its default when resetting the
i210 to resolve this issue.

Signed-off-by: Jeff Westfahl <jeff.westfahl@ni.com>

NatInst-CAR-ID: 470113
Natinst-ReviewBoard-ID: 67009