igb: setup link after powering up the PHY
We've been investigating a failure to link of the i210 with a specific
DLink switch, the DGS-1016D. The failure to link occurs only with the
in-tree igb driver, the out-of-tree igb driver links correctly.

I instrumented the in-tree and out-of-tree igb drivers and identified
a difference in behavior where the out-of-tree driver calls an internal
link setup function as part of the PHY power-up process. The in-tree
driver has a similar link setup function, but does not call it when the
PHY is powered up. Adding a call to the existing link setup function to
the in-tree driver to match the out-of-tree driver resolves the link
failure problem with the DLink switch.

Signed-off-by: Jeff Westfahl <jeff.westfahl@ni.com>

NatInst-CAR-ID: 470056
Natinst-ReviewBoard-ID: 68944