usb: remove unnecessary USB_OTG dependency on power management
The USB gadget support currently depends on power management
(CONFIG_PM_RUNTIME) being enabled, but we do not enable power management
on Smasher.  We get around this by hardcoding CONFIG_USB_OTG=y in our
defconfig, and a quick look at the code doesn't turn up any apparent
reason for this dependency.  Removing this dependency to eliminate the
chance of someone breaking the USB gadget support by doing 'make
nati_x86_64_defconfig' followed by 'make savedefconfig'.

Note that this was broken on the Smasher 14.0/3.10 kernel branch, but
seems to be OK on the unified 3.10 branch.

Signed-off-by: Ben Shelton <ben.shelton@ni.com>