xilinx_emacps: Add support for fpga_clk
The Zynq microcontroller allows Ethernet to be clocked from either EMIO
or MIO. This option is generally chosen by the bootloader by setting or
clearing the SRCSEL bit in the Zynq SLCR's GEM0_RCLK_CTRL or
GEM1_RCLK_CTRL register. As currently written, the xilinx_emacps driver
does not attempt to configure divisors or ungate the EMIO-sourced clock:
fclk0, fclk1, fclk2, or fclk3, as determined by the programmed bitfile.
It is necessary that if the Zynq Ethernet should be clocked by EMIO that
the xilinx_emacps driver control the clock for different link speeds.

At a minimum, the "fpga_clk" must be specified in the device tree. This
is to give the xilinx_emacps driver enough knowledge to know which clock
to control for different link speeds. We now check for the presence of
an "fpga_clk" specified in the device tree for the Ethernet peripheral.

We also now, if an fpga_clk is specified, control that clock (divisors
and gate) when the link changes or when the Ethernet peripheral is
suspended or resumed.

It is okay to still change and ungate the gem_clk, even if SRCSEL bit in
the RCLK is set. No harm comes from configuring the gem_clk divisors
despite it not being used. If fact, it is necessary that the gem_clk be
ungated (i.e. enabled) in order for the EMIO-sourced clock to be
ungated; this is why we take the approach of preparing, configuring, and
enabling both the fpga_clk and the gem_clk.

Signed-off-by: Jaeden Amero <jaeden.amero@ni.com>