xilinx_emacps: Use the MII ID of the PHY parent
Previously, the MII bus ID of the PHY grandparent was being used. This
would erroneously cause an ID of zero to be used for both MII buses in
the case where the board-specific device tree requests two distinct MII
buses. By use the MII bus ID of the PHY parent instead, the MII ID
specified in the device tree for the Ethernet interface is adhered to.

Signed-off-by: Jaeden Amero <jaeden.amero@ni.com>