ni-tecate: Comment about Ethernet FPGA clock
For Dos Equis boards, we had to specify the FPGA clock in order for
secondary Ethernet to work. For Tecate boards, specifying the FPGA clock
will cause secondary Ethernet to not work. For the sake of being
obvious, we add comments to the Ethernet interface sections of Tecate
commenting on why we don't specify the FPGA clock for either of the
Ethernet interfaces.

Signed-off-by: Jaeden Amero <jaeden.amero@ni.com>