ni-tecate: Use two separate MII buses
For Tecate boards, primary and secondary Ethernet do not share one MII
bus as on Dos Equis boards. As such, we need the secondary Ethernet PHY
to have its own MII bus interface created by the driver. To allow the
creation of this bus, we remove "xlnx,no_mdio_bus" from the Ethernet
interface description in the device tree.

Signed-off-by: Jaeden Amero <jaeden.amero@ni.com>