zynq-7000: Use the Zynq-specific pl310
We need to specify the Zynq-specific pl310, "xlnx,ps7-pl310-1.00.a", and
not the generic one, "arm,pl310-cache", so that Zynq-specific setup code
will run before the initialization of the L2 cache. This is necessary to
avoid L2 cache corruption on at least a few Zynq boards based on the
zynq-7000 device tree.

Signed-off-by: Jaeden Amero <jaeden.amero@ni.com>
Reviewed-by: Joe Hershberger <joe.hershberger@ni.com>
Reviewed-by: Gratian Crisan <gratian.crisan@ni.com>
Reviewed-by: Ben Shelton <ben.shelton@ni.com>