ARM: Zynq: l2x0: Avoid cache corruption
The Xilinx Zynq CPU has a bug where the L2 cache will return invalid
data in some circumstances unless the Zynq L2C_RAM register is set to
0x00020202 before the first enabling of the L2 cache. At least a few
Zynq-based boards exist with bootloaders that do not properly set the
L2C_RAM register before enabling the L2 cache. A subset of these boards
first enables the L2 cache in Linux; i.e. the code that runs before
Linux on these does not enable the L2 cache.

In order to ensure that the L2C_RAM register is set to 0x00020202 before
the L2 cache is enabled on the aforementioned subset of boards, we
provide Zynq-specific L2 cache setup code that runs before the first
enablement of the L2 cache on said boards.

For boards that have bootloaders that do set the L2C_RAM register
correctly, there is no harm in setting this value again. As a fallback,
they can specify compatibility with "arm,pl310-cache" instead of
"xlnx,ps7-pl310-1.00.a" in their board device trees, as currently the
only difference with "arm,pl310-cache" behavior is the initialization of
L2C_RAM.

The Zynq bug is described by Xilinx AR# 54190 as quoted below.

Xilinx AR# 54190
http://www.xilinx.com/support/answers/54190.htm
Captured on 2014-09-24 14:43 -0500

  = Description =
  For proper L2 cache operation, the user code must program the
  slcr.L2C_RAM register (address 0xF800_0A1C) to the value of
  0x0002_0202 before enabling the L2 cache. The reset value
  (0x0001_0101) might cause, very infrequently, the L2 cache to return
  invalid data.

  = Solution =
  It is up to the user code (FSBL or other user code) to set the
  slcr.L2C_RAM register to the value 0x0002_0202 before enabling the L2
  cache.

  Note: The L2 cache is disabled after reset and is not enabled by the
  BootROM.

  Note: The slcr.l2C_RAM register was previously reserved. It is added
  in the Zynq-7000 AP SoC Technical Reference Manual (TRM) v1.5 as
  "Reserved".

Signed-off-by: Jaeden Amero <jaeden.amero@ni.com>
Reviewed-by: Joe Hershberger <joe.hershberger@ni.com>
Reviewed-by: Gratian Crisan <gratian.crisan@ni.com>
Reviewed-by: Ben Shelton <ben.shelton@ni.com>