ni-solbetter: Specify the gem0 and gem1 FPGA clock
Specify that fclk0 should be used as the FPGA clock for primary Ethernet
(gem0) and that fclk2 should be used as the FPGA clock for secondary
Ethernet (gem1). This communicates to the xilinx_emacps driver to
control fclk0 and fclk1 when the link changes or when the Ethernet
peripheral suspends or resumes.

Signed-off-by: Jaeden Amero <jaeden.amero@ni.com>
Acked-by: Ben Shelton <ben.shelton@ni.com>
Acked-by: Gratian Crisan <gratian.crisan@ni.com>
Acked-by: Keng Soon Cheah <keng.soon.cheah@ni.com>
---
Natinst-ReviewBoard-ID: 79772