ni-tecate: Enable FPGA clocks
We need to enable all FPGA clocks that don't have a driver associated
with them. This is to prevent the common clock framework from disabling
any parent clocks of the FPGA clocks as well as the FPGA clocks
themselves. For Tecate boards, we enable fclk0 for primary and secondary
Ethernet (which is used as a constant 125 MHz clock) and fclk1 for
serial.

Signed-off-by: Jaeden Amero <jaeden.amero@ni.com>