Revert "nizynq: don't disable fclk[0-3] as unused"
This commit accidentally reverted some "enable_reg" code. It shouldn't
have done that. This was probably due to a merge error when moving this
commit from 3.10 to 3.14.

In addition, Xilinx upstream has a different way to resolve the problem
that this commit originally intended to solve that I like better. We'll
revert this commit and use the upstream solution instead.

This reverts commit 1a6d2cac78ee919c899276f8cbe0a58a9d18e800.

Signed-off-by: Jaeden Amero <jaeden.amero@ni.com>