net: xilinx_emacps: Add support for device polling
Add support for the device_poll and netdev_poll interfaces to the Xilinx
Zynq Ethernet driver.  This allows Zynq devices to be used without
generating any hardware interrupts.  This support was patterned after
the reference implementation in the Intel igb driver in the commit 'igb:
support device_poll and netdev_poll interfaces'.

Due to the complexities of enabling and disabling interrupts in this
driver, poll continuously, rather than hooking into NAPI and disabling
polling while bottom-half packet handling is in progress.

Signed-off-by: Ben Shelton <ben.shelton@ni.com>
Acked-by: Jaeden Amero <jaeden.amero@ni.com>
---
Natinst-ReviewBoard-ID: 79822