Revert "Revert "ARM: Zynq: l2x0: Avoid cache corruption""
This reverted commit inadvertently disabled the L2 cache by not also
changing the compatible string of the zynq-7000.dtsi back to
"arm,pl310-cache". This means that the OF matching was looking for a
compatible string of "xlnx,ps7-pl310-1.00.a" instead of
"arm,pl310-cache". It didn't find any drivers that were compatible with
it, so the L2 cache never got enabled.

The reason why reverting the original commit fixed the DMA streaming
tests is because it disabled the L2 cache on accident, not because of
any bugs in the original commit nor due to its setting of L2C_RAM.

This reverts commit e30273961e4af261eb631427a9f7c557001cfd8e.

Signed-off-by: Jaeden Amero <jaeden.amero@ni.com>
Acked-by: Joshua Hernstrom <joshua.hernstrom@ni.com>
Acked-by: Ben Shelton <ben.shelton@ni.com>
Acked-by: Gratian Crisan <gratian.crisan@ni.com>
Acked-by: Brad Mouring <brad.mouring@ni.com>
---
Natinst-CAR-ID: 496676
Natinst-ReviewBoard-ID: 80140