fixup: ARM: Zynq: l2x0: Avoid cache corruption
This is a fixup to e9e1f08e8216535ce3dc39381d6a16cf52e1cd11 and should
be applied at first opportunity. When it is applied, the
"Natinst-CAR-ID: 411111" tag should be added to the original commit
description, if that tag is still desired.

We previously used an incorrect mask of 0x30303 instead of 0x70707. This
meant that if any of the extra bits were populated, they wouldn't get
set to zero. This wasn't ever a problem in practice because whenever
this code was run, the code the L2C_RAM register was always at its
power-on default value of 0x10101.

We also fixed a mistake in a comment.

Signed-off-by: Jaeden Amero <jaeden.amero@ni.com>
Acked-by: Joshua Hernstrom <joshua.hernstrom@ni.com>
Acked-by: Ben Shelton <ben.shelton@ni.com>
Acked-by: Gratian Crisan <gratian.crisan@ni.com>
Acked-by: Brad Mouring <brad.mouring@ni.com>
---
Natinst-CAR-ID: 411111
Natinst-ReviewBoard-ID: 80140