nati_zynq_defconfig: Move USB Ethernet gadget support to built-in
On the 3.2 kernel, CONFIG_USB_ETH was set to built-in, but in the
process of migrating to the 3.10 kernel, it was moved to building as a
module.  This causes the USB Ethernet gadget support not to be loaded at
boot time, which means that the target can't be connected to.  To fix
this issue, move this support back to built-in.

Signed-off-by: Ben Shelton <ben.shelton@ni.com>
Acked-by: Gratian Crisan <gratian.crisan@ni.com>
---
Natinst-CAR-ID: 496659
Natinst-ReviewBoard-ID: 80309