HACK mach-zynq: Don't ignore sharable attribute
This is a big hack with serious consequences. It must never be shipped
to customers. It must never be upstreamed. It should be reverted at
first opportunity.

ARM and Xilinx created a fix for a data corruption issue they
encountered with Linux-native DMA. This fix makes sense and we want to
ship with it. However, it has the consequence of breaking how NI-RIO
currently does DMA between the FPGA and CPU (i.e. RIO Host Interface
DMA).

The Xilinx and ARM fixes are implemented in the following commits.

From ARM:
1a8e41cd672f894bbd74874eac601e6cedf838fb

From Xilinx, copying ARM for their Zynq:
1a85939af40acca2bf963407b497cc31c303ff3e

For the sake of lowering the priority of the P0 stackbuster CAR, CAR
496676, and for the lack of a better short term solution, we will
temporarily not set the L2 cache configuration bit to ignore the
sharable attribute (Zynq TRM UG585 - reg1_aux_controls's
shared_attr_override_en).

Signed-off-by: Jaeden Amero <jaeden.amero@ni.com>
Acked-by: Gratian Crisan <gratian.crisan@ni.com>
Acked-by: Joshua Hernstrom <joshua.hernstrom@ni.com>
Acked-by: Kyle Teske <kyle.teske@ni.com>
Acked-by: Xander Huff <xander.huff@ni.com>
--
Natinst-CAR-ID: 496676
Natinst-ReviewBoard-ID: 80865