nati_x86_64_defconfig: Move USB Ethernet gadget support to built-in
On the 3.10 kernel, CONFIG_USB_ETH was set to built-in, but in the
process of migrating to the 3.14 kernel, it was moved to building as a
module. This causes the USB Ethernet gadget support not to be loaded at
boot time. To fix this issue, move this support back to built-in.

Signed-off-by: Joshua Hernstrom <joshua.hernstrom@ni.com>
Acked-by: Gratian Crisan <gratian.crisan@ni.com>
---
Natinst-ReviewBoard-ID: 80877