nati_zynq_defconfig: Offload RCU callbacks to kthreads
Enable CONFIG_RCU_NOCB_CPU option which offloads the RCU callback processing
to kernel threads. A set of threads named rcuo<x>/<N> are created.
"x" is "b" for RCU-bh, "p" for RCU-preempt, and "s" for RCU-sched. <N> is the
CPU number. These threads can be pinned to a CPU via cgroups.
This allows a user to reserve a CPU for high priority real-time workloads.

Also enable CONFIG_RCU_NOCB_CPU_ALL which specifies that all CPUs are
no callback CPUs when it comes to RCU. Essentially the RCU threads described
above are created for all CPUs (not just the ones specified on the kernel
command line). This allows for greater flexibility in configuring the core
affinity of these kernel threads from user-space.

Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>
Acked-by: Ben Shelton <ben.shelton@ni.com>
Acked-by: Brad Mouring <brad.mouring@ni.com>