fixup: xemacps: Ensure we unmap bad DMA buffers
This is a fixup to 1c145c7c5c298252f3d663a9e6ac8b7bd023fa69 and should
be applied at first opportunity.

After mapping DMA buffers, we calculate the length and compare to the
maximum buffer size and the minimum Ethernet packet length. If this
calculated length falls outside the range, we skip the packet.
However, we need to make sure we also unmap the buffer we've already
mapped before we skip it.

Signed-off-by: Xander Huff <xander.huff@ni.com>
Acked-by: Ben Shelton <ben.shelton@ni.com>
Acked-by: Jaeden Amero <jaeden.amero@ni.com>
---
Natinst-ReviewBoard-ID: 82259