phy: Run the PHY state machine in start and stop
This is a squash for fd456445dcb07e9009a4a275ab732cc11295aeb7, "phy: add
PHY_NO_UNNECESSARY_POLLING config option". It should get squished with
that commit, with some of the notes here added to the new resulting
commit message.

When resuming a PHY, interrupts will be disabled. This means that the
PHY state machine will never run when PHY_NO_UNNECESSARY_POLLING is set,
as it will only run the PHY state machine based on interrupt actions or
when the PHY is first connected. This results in the PHY state machine
getting stuck in PHY_RESUMING. To resolve this, we run the state machine
in phy_start, which, for PHY_RESUMING, will reenable the PHY interrupt
and handle any pending phy changes from before interrupts were enabled.

Since PHY_NO_UNNECESSARY_POLLING will stop polling the state machine
after one shot, which may not enable interrupts if the state isn't
PHY_RESUMING, we also enable interrupts in phy_start after the call to
phy_start_machine.

In addition, if the PHY begins autonegotiation and enters the PHY_AN
state, the state machine will poll once after phy_start without enabling
interrupts. Turn interrupts on in phy_start to avoid this.

When stopping a PHY, we also want the adjust_link handler to get called.
This happens from the PHY_HALTED state. So, we run the machine in
phy_stop, as well.

Signed-off-by: Nathan Sullivan <nathan.sullivan@ni.com>
Signed-off-by: Jaeden Amero <jaeden.amero@ni.com>
Acked-by: Ben Shelton <ben.shelton@ni.com>
Acked-by: Gratian Crisan <gratian.crisan@ni.com>
---
Natinst-ReviewBoard-ID: 84558