Revert "HACK mach-zynq: Don't ignore sharable attribute"
This reverts commit 1d8e395ac5a3560f55b807c7c46cdc5b52a06873.

We have fixed the RIO driver so that it does DMA correctly by changing
the AXI AxCACHE settings to 0x6 (write through, allocate on write). With
the L2 controller ignoring the shared line, the previous setting of 0x2
would not update lines in the L2 and lead to memory aliasing.

As a bonus, all this USB watchdog timer, ACCESS_ONCE ugliness, and
reading back of memory (which we did to cope with the original bug that
ignoring the sharable attribute solves) can go away.

Signed-off-by: Jaeden Amero <jaeden.amero@ni.com>
Acked-by: Nathan Sullivan <nathan.sullivan@ni.com>
Acked-by: Ben Shelton <ben.shelton@ni.com>
---
Natinst-ReviewBoard-ID: 86910