lci gadget: Use 'product' param for USB descriptor and mass storage
We supply the product name as a string on the kernel command line,
in a form such as "NI VB-8012". (It actually bubbles out from u-boot,
as the "DeviceDesc" environment variable.)

Previously we were only using this to populate g_lci.product, which
is used to set the product string descriptor. We had g_lci.fsg_prod
as a separate tunable for the mass storage product name; instead,
remove it and derive the mass storage name from g_lci.product.

Signed-off-by: Brandon Streiff <brandon.streiff@ni.com>
Acked-by: William Earle <william.earle@ni.com>
Acked-by: Zach Hindes <zach.hindes@ni.com>
--
Natinst-CAR-ID: 513062
Natinst-ReviewBoard-ID: 88772