defconfig, zynq, x86_64: Disable dynamic ticks on idle
Disable dynamic ticks on idle. The option has a >10uS latency hit on ARM.
The main use for dynamic ticks on idle is for power management where on an
otherwise idle system the periodic tick will keep the system from going into
deeper sleep states. However this makes the transitions from idle to an RT
thread more costly and we are currently not doing any power management on our
controllers.

We will revisit this in the future when full dynamic ticks support is better
tested and considered stable/safe for use.

Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>
Acked-by: Ben Shelton <ben.shelton@ni.com>
Acked-by: Scot Salmon <scot.salmon@ni.com>
---
Natinst-ReviewBoard-ID: 88882