nizynqcpld: Strobe the watchdog control register
The CPLD watchdog control register is a strobe register, full of a mix
of read-only and write-only bits. Because it is a strobe register, this
driver was in error doing read, modify, write when updating the watchdog
control register. Some bad things happened because of the driver's
misuse of this register.

A specific example of a failure due to read, modify, write of the
watchdog control register follows. Bit 7 of the watchdog control
register is "Watchdog Mode". A value of 1 here means "Boot Mode". A
value of 0 means "User Mode". Bit 4 ("Watchdog Pet") is a write only bit
that pets the watchdog when set to 1. When attempting to pet the
watchdog while the watchdog is in boot mode, the driver would do a read,
see that bit 7 is set (indicating that the watchdog is in boot mode), or
this value with bit 4 (0x10) to say "I want to pet the boot mode watch
dog," and end up writing both bit 7 and bit 4 as 1. Writing a 1 to bit 7
means "I want the watchdog to enter user mode," so no petting occurs,
despite the driver's intent to pet the watch dog and not to enter user
mode.

Signed-off-by: Jaeden Amero <jaeden.amero@ni.com>
Reviewed-by: Ben Shelton <ben.shelton@ni.com>
Reviewed-by: Gratian Crisan <gratian.crisan@ni.com>
---
Natinst-ReviewBoard-ID: 90484