fixup! nizynqcpld: clean up tecate LEDs
We accidentally removed eth1 from the Sol LEDs. Add them back to Sol and
remove them from Tecate.

Tecate uses a CPLD with a Product Code of 2 and Sol uses a CPLD with a
Product Code of 3. The mistaken commit changed the table for CPLDs with
Product Codes of 3, so we know we really did make an accident and
changed the Sol LEDs when we actually meant to change the Tecate LEDs.

Signed-off-by: Jaeden Amero <jaeden.amero@ni.com>
Reviewed-by: Ben Shelton <ben.shelton@ni.com>
Reviewed-by: Josh Cartwright <joshc@ni.com>
---
Natinst-ReviewBoard-ID: 91284