HACK net/phy: Don't use a NULL attached_dev
Since local commit 9f2ba7de76e2, it's possible that the phy's attached_dev
is NULL after phy_stop is called (e.g. via device_release_driver via
EtherCAT) on a device that uses a PHY on an MDIO bus it doesn't own. Check
for NULL to make sure we don't crash.

This is a hack because the PHY is linked, but doesn't have an attached_dev.
This seems like it shouldn't happen. How are we ending up linked without an
attached_dev?

Signed-off-by: Jaeden Amero <jaeden.amero@ni.com>
Reviewed-by: Xander Huff <xander.huff@ni.com>
---
Natinst-CAR-ID: 528734
Natinst-ReviewBoard-ID: 100230