cfg80211: wext: Force scans to occur in AP mode
When in AP mode, we should force scans to happen from the wext
compatibility layer.  This won't hurt when in station mode, and
only affects the wl12xx driver.

This change is to retain parity with NI's shipped 3.2 kernel.  It
can be dropped when we switch away from using the WEXT interface,
and should not be upstreamed.

Signed-off-by: James Minor <james.minor@ni.com>
Reviewed-by: Ben Shelton <ben.shelton@ni.com>
Reviewed-by: Jaeden Amero <jaeden.amero@ni.com>
Natinst-ReviewBoard-ID: 92027
[gratian: fix trivial conflict with f0df91b6a712 ("wifi: cfg80211: hide scan internals")]
Signed-off-by: Gratian Crisan <gratian.crisan@emerson.com>