UBIFS: Fix up assertion for 3.14 kernel
Reverting the patch 'UBIFS: fix a couple bugs in UBIFS xattr length
calculation' and cherry-picking the upstream version caused a regression
where an ubifs_assert() in ubifs_jnl_update() was triggered repeatedly.

To fix this issue, only check the assert in the non-extended-attribute
case. In Linux 4.0, where the upstream patch was cherry-picked from,
this ubifs_assert() is not present, so this check is not needed.

Fixes: a05447ae ("UBIFS: fix a couple bugs in UBIFS xattr length calculation")
Signed-off-by: Ben Shelton <ben.shelton@ni.com>
Reviewed-by: Brad Mouring <brad.mouring@ni.com>
Reviewed-by: Gratian Crisan <gratian.crisan@ni.com>
---
Natinst-ReviewBoard-ID: 93749