sched: Disable smarter wake-affine logic
The commit 'sched: Implement smarter wake-affine logic' introduced a
scheduler change that benefits some workloads but reduces the NI
RT-FIFO's maximum throughput.  Disable the wake_wide functionality
introduced in this commit.  On x64 targets, this brings performance back
to previous levels.  On Zynq targets, it improves CPU utilization, but
to get the highest possible throughput for applications that are
throughput-bound and not compute-bound, we will recommend that customers
pin the RT-FIFO writer thread and reader thread to the same core.

This change is to retain parity with NI's shipped 3.2 (Zynq) and 3.10
(x64) kernels, so it should not be upstreamed.

Reported-by: Nathan Sullivan <nathan.sullivan@ni.com>
Signed-off-by: Ben Shelton <ben.shelton@ni.com>
Reviewed-by: Jaeden Amero <jaeden.amero@ni.com>
---
Natinst-CAR-ID: 512421
Natinst-ReviewBoard-ID: 93987