Revert "macb: Don't create MDIO bus based on device tree"
This reverts commit b07f131c7ea392ba9361819383aa7c9a57675ade.

We didn't need it. It is actually doesn't hurt anything that an MDIO bus
is created.

Even if we really wanted to prevent MDIO buses from getting created, we
didn't need to add a new device tree attribute for this. We could have
seen that the MDIO bus didn't need to exist if the mii_bus has no PHY
children after macb_mii_init and cleaned up the bus at such a time.

Signed-off-by: Jaeden Amero <jaeden.amero@ni.com>
Fixes: 9f2ba7de76e2 ("macb: Use the PHY via PHY handle")
Reviewed-by: Ben Shelton <ben.shelton@ni.com>
Reviewed-by: Josh Cartwright <joshc@ni.com>
---
Natinst-CAR-ID: 517548
Natinst-ReviewBoard-ID: 95573